added all crud functionalities in the patient controller + updated patient DAO for the crud + fixed patient forms and added new ones

+++ src/main/java/com/mycompany/dao/PatientDAO.java
@@ -21,6 +21,15 @@ public class PatientDAO {
         }
     }
     
+    public Patient getPatientById(int id) {
+        try (Session session = HibernateUtil.getSessionFactory().openSession()) {
+            return session.get(Patient.class, id);
+        } catch (Exception e) {
+            e.printStackTrace();
+            return null;
+        }
+    }
+    
     public boolean createPatient(Patient patient) {
         Transaction transaction = null;
         try (Session session = HibernateUtil.getSessionFactory().openSession()) {
@@ -66,24 +75,45 @@ public class PatientDAO {
             return false;
         }
     }
-
-    public boolean deletePatient(Patient patient) {
+    public void deletePatientById(int id) {
         Transaction transaction = null;
         try (Session session = HibernateUtil.getSessionFactory().openSession()) {
+            // Begin a transaction
             transaction = session.beginTransaction();
-            
-            session.delete(patient);
-            
-            transaction.commit();
-            return true;
+
+            // Retrieve the patient by ID
+            Patient patient = session.get(Patient.class, id);
+
+            // If the patient exists, delete it
+            if (patient != null) {
+                session.delete(patient);
+                // Commit the transaction
+                transaction.commit();
+            }
         } catch (Exception e) {
             if (transaction != null) {
                 transaction.rollback();
             }
             e.printStackTrace();
-            return false;
         }
     }
+//    public boolean deletePatient(Patient patient) {
+//        Transaction transaction = null;
+//        try (Session session = HibernateUtil.getSessionFactory().openSession()) {
+//            transaction = session.beginTransaction();
+//            
+//            session.delete(patient);
+//            
+//            transaction.commit();
+//            return true;
+//        } catch (Exception e) {
+//            if (transaction != null) {
+//                transaction.rollback();
+//            }
+//            e.printStackTrace();
+//            return false;
+//        }
+//    }
     
     public boolean updatePatientPassword(Patient patient, String newPassword) {
         Transaction transaction = null;
